+++ lib/versao.ts
@@ -1,3 +1,3 @@
-export const VERSAO = '0.3.1';
+export const VERSAO = '0.3.2';
 
-export const VERSAO_NOME = 'No ar';
+export const VERSAO_NOME = 'Números de verdade';